+++ yarn-project/ivc-integration/src/wasm_client_ivc_integration.test.ts
@@ -1,11 +1,16 @@
+import { BB_RESULT, verifyClientIvcProof, writeClientIVCProofToOutputDirectory } from '@aztec/bb-prover';
 import { AztecClientBackend } from '@aztec/bb.js';
+import { createLogger } from '@aztec/foundation/log';
 
 import { jest } from '@jest/globals';
 
 /* eslint-disable camelcase */
 import createDebug from 'debug';
 import { ungzip } from 'pako';
+import path from 'path';
+import { fileURLToPath } from 'url';
 
+import { getWorkingDirectory } from './bb_working_directory.js';
 import {
   MOCK_MAX_COMMITMENTS_PER_TX,
   MockAppCreatorCircuit,
@@ -19,6 +24,7 @@ import {
   MockPrivateKernelResetCircuit,
   MockPrivateKernelResetVk,
   MockPrivateKernelTailCircuit,
+  generate3FunctionTestingIVCStack,
   getVkAsFields,
   witnessGenCreatorAppMockCircuit,
   witnessGenMockPrivateKernelInitCircuit,
@@ -27,9 +33,11 @@ import {
   witnessGenMockPrivateKernelTailCircuit,
   witnessGenReaderAppMockCircuit,
 } from './index.js';
-import { proveThenVerifyAztecClient } from './prove_wasm.js';
+import { proveClientIVC as proveClientIVCNative } from './prove_native.js';
+import { proveClientIVC as proveClientIVCWasm, proveThenVerifyAztecClient } from './prove_wasm.js';
 
-const logger = createDebug('ivc-integration:test:wasm');
+const logger = createLogger('ivc-integration:test:wasm');
+
 createDebug.enable('*');
 
 jest.setTimeout(120_000);
@@ -42,40 +50,42 @@ describe('Client IVC Integration', () => {
   // 2. Run the init kernel to process the app run
   // 3. Run the tail kernel to finish the client IVC chain.
   it('Should generate a verifiable client IVC proof from a simple mock tx via bb.js', async () => {
-    const tx = {
-      number_of_calls: '0x1',
-    };
-    // Witness gen app and kernels
-    const appWitnessGenResult = await witnessGenCreatorAppMockCircuit({ commitments_to_create: ['0x1', '0x2'] });
-    logger('generated app mock circuit witness');
-
-    const initWitnessGenResult = await witnessGenMockPrivateKernelInitCircuit({
-      app_inputs: appWitnessGenResult.publicInputs,
-      tx,
-      app_vk: getVkAsFields(MockAppCreatorVk),
-    });
-    logger('generated mock private kernel init witness');
-
-    const tailWitnessGenResult = await witnessGenMockPrivateKernelTailCircuit({
-      prev_kernel_public_inputs: initWitnessGenResult.publicInputs,
-      kernel_vk: getVkAsFields(MockPrivateKernelInitVk),
-    });
-    logger('generated mock private kernel tail witness');
-
-    // Create client IVC proof
-    const bytecodes = [
-      MockAppCreatorCircuit.bytecode,
-      MockPrivateKernelInitCircuit.bytecode,
-      MockPrivateKernelTailCircuit.bytecode,
-    ];
-
-    logger('built bytecode array');
-    const witnessStack = [appWitnessGenResult.witness, initWitnessGenResult.witness, tailWitnessGenResult.witness];
-    logger('built witness stack');
+    const [bytecodes, witnessStack] = await generate3FunctionTestingIVCStack();
 
     const verifyResult = await proveThenVerifyAztecClient(bytecodes, witnessStack);
-    logger(`generated then verified proof. result: ${verifyResult}`);
+    logger.info(`generated then verified proof. result: ${verifyResult}`);
 
+    // We use the bb binary for verification / writing out the VK
+    const bbBinaryPath = path.join(
+      path.dirname(fileURLToPath(import.meta.url)),
+      '../../../barretenberg/cpp/build/bin',
+      'bb',
+    );
+    const clientIVCWorkingDirectory = await getWorkingDirectory('bb-client-ivc-integration-');
+    const tasks = [
+      proveClientIVCNative(bbBinaryPath, clientIVCWorkingDirectory, witnessStack, bytecodes, logger),
+      proveClientIVCWasm(bytecodes, witnessStack),
+    ];
+    const [_, wasmProof] = await Promise.all(tasks);
+
+    // Verify the clientIVCWorkingDirectory, written to by native proveClientIVC
+    const verifyNativeResult = await verifyClientIvcProof(
+      bbBinaryPath,
+      clientIVCWorkingDirectory.concat('/proof'),
+      clientIVCWorkingDirectory.concat('/vk'),
+      logger.info,
+    );
+    expect(verifyNativeResult.status).toEqual(BB_RESULT.SUCCESS);
+
+    // Write the WASM proof over the output directory (the bb cli will have output to this folder, we need the vk to be in place).
+    await writeClientIVCProofToOutputDirectory(wasmProof, clientIVCWorkingDirectory);
+    const verifyWasmResult = await verifyClientIvcProof(
+      bbBinaryPath,
+      clientIVCWorkingDirectory.concat('/proof'),
+      clientIVCWorkingDirectory.concat('/vk'),
+      logger.info,
+    );
+    expect(verifyWasmResult.status).toEqual(BB_RESULT.SUCCESS);
     expect(verifyResult).toEqual(true);
   });
 
@@ -93,7 +103,7 @@ describe('Client IVC Integration', () => {
     // Compute the numbers of gates in each circuit
     const gateNumbers = await backend.gates();
     await backend.destroy();
-    logger('Gate numbers for each circuit:', gateNumbers);
+    logger.info('Gate numbers for each circuit:', gateNumbers);
     // STARTER: add a test here instantiate an AztecClientBackend with the above bytecodes, call gates, and check they're correct (maybe just
     // eyeball against logs to start... better is to make another test that actually pins the sizes since the mock protocol circuits are
     // intended not to change, though for sure there will be some friction, and such test should actually just be located in barretenberg/ts)
@@ -160,7 +170,7 @@ describe('Client IVC Integration', () => {
     ];
 
     const verifyResult = await proveThenVerifyAztecClient(bytecodes, witnessStack);
-    logger(`generated then verified proof. result: ${verifyResult}`);
+    logger.info(`generated then verified proof. result: ${verifyResult}`);
 
     expect(verifyResult).toEqual(true);
   });
